Add span context tracking metrics
Both stages of the feature now report, each under its own metric, tagged
with the tracer language:

  span_tracking.resolve.{error,success}  once per traced process, working
                                         out how to read it
  span_tracking.read.{error,success}     once per event, capturing what
                                         the readers found

They are counted apart because they answer different questions, over
denominators orders of magnitude apart.

Resolving happens in user space, and the reason travels with the error:
unsupported_version is the hand-maintained Go offset table falling behind
a release, bad_symbol points at the tracer rather than at us, map_put is a
full map. Reading happens in the probes, which now say why they gave up
through an error field on span_context_t.

Only genuine failures are counted. A tracer that does not implement OTEP
4947 exports no otel_thread_ctx_v1, a goroutine may hold no pprof labels,
and a thread may not be inside a span at all: these are the expected
outcome and are not counted as errors.

+++ pkg/security/resolvers/otelattrs/resolver.go
@@ -52,15 +52,15 @@ type Attribute struct {
 // slot. The kernel stamps each snapshot with its id, so a mismatch means the slot
 // was recycled for a later snapshot before this event got drained, and the
 // attributes are gone.
-func (r *Resolver) Resolve(attrsID uint64) ([]Attribute, error) {
-	key := uint32(attrsID % maxEntries)
+func (r *Resolver) Resolve(attrsID uint32) ([]Attribute, error) {
+	key := attrsID % maxEntries
 
 	var entry kernelAttrsEntry
 	if err := r.attrsMap.Lookup(key, &entry); err != nil {
 		return nil, fmt.Errorf("unable to resolve the otel span attributes for `%d`: %w", attrsID, err)
 	}
 
-	if uint64(entry.ID) != attrsID {
+	if entry.ID != attrsID {
 		return nil, fmt.Errorf("incorrect id `%d` vs `%d`", attrsID, entry.ID)
 	}
